Resolve #14 - renamed errors package
The `errors` package has been renamed to be `hterror` so as to make it
play more nicely with the stdlib.

All imports have been fixed.
