+++ hterror/errors.go
@@ -1,4 +1,4 @@
-package errors
+package hterror
 
 import (
 	"fmt"
